Speed up (my) RSA algorithm
I originally used gen_safe_prime() to generate the p and q values,
because not using it caused d to be None a good bit of the time, but
that was really slow. Instead, I'm using gen_prime, and just retrying if
d comes out as None. Also, changed the key modulus and exponent to be
BigUint rather than BigInt.